modified the service and controller class to mae the service loose couple and updated unit test

+++ src/main/java/com/acoustic/repository/TaxRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TaxRepository extends JpaRepository<Tax, Long> {
+public interface TaxRepository extends JpaRepository<Tax, Integer> {
 
 
 }
